% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_ROI_list.R
\name{Cluster_ROI_list}
\alias{Cluster_ROI_list}
\title{Cluster ROI Data from a List of Regions}
\usage{
Cluster_ROI_list(roi_list, allow_further_itteration = FALSE)
}
\arguments{
\item{roi_list}{A list of datasets, each representing a Region of Interest (ROI), to be clustered.}

\item{allow_further_itteration}{Logical; if \code{TRUE}, the clustering process may perform additional iterative refinement.}
}
\value{
A list of clustered ROI data, excluding any entries that failed due to missing profiles or errors.
}
\description{
This function performs clustering on a list of Regions of Interest (ROI) using a targeted profile clustering method.
It optionally allows further iteration during clustering and removes any clusters that contain error messages, such as missing profiles.
}
\details{
The function internally uses \code{targeted_profile_clusterer()} to cluster the input ROI datasets.
It checks for and removes any cluster outputs that contain error messages like "no profiles" to ensure the returned list only includes valid clusters.
}
\examples{
\dontrun{
rois <- list(roi1, roi2, roi3)
clusters <- Cluster_ROI_list(rois, allow_further_itteration = TRUE)
}

}
\seealso{
\code{\link{targeted_profile_clusterer}}
}
