% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_build_graphs_for_bimodal_regions.R
\name{comparisonplotbuilder3}
\alias{comparisonplotbuilder3}
\title{Build Graphs for comparison Bimodal regions}
\usage{
comparisonplotbuilder3(
  clusters,
  originaldata,
  angle_data,
  diameter_data,
  radius_data,
  umaplist,
  selected_datasets,
  miniumapgraphs,
  Cluster_consensus_images,
  table_list
)
}
\description{
Build Graphs for comparison Bimodal regions
}
