% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extreme_angle_detector.R
\name{Extreme_angle_detector}
\alias{Extreme_angle_detector}
\title{Extreme angle detector
This tags the nuclei with angles over 280 that are likely artifacts}
\usage{
Extreme_angle_detector(data)
}
\arguments{
\item{data}{the nuclear measurements exported dataset from NMA}
}
\description{
Extreme angle detector
This tags the nuclei with angles over 280 that are likely artifacts
}
