% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graphs_for_whole_dataset.R
\name{fulldatasetclustergrapher}
\alias{fulldatasetclustergrapher}
\title{Build graphs for whole dataset tab}
\usage{
fulldatasetclustergrapher(data, umaplist)
}
\description{
Build graphs for whole dataset tab
}
