% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartigansdipper.R
\name{get.dip.test.regions}
\alias{get.dip.test.regions}
\title{Get columns of a data.frame that are non-unimodal}
\usage{
get.dip.test.regions(data, dip.test.alpha = 0.05, is.profile = TRUE)
}
\arguments{
\item{data}{the data frame to input}

\item{dip.test.alpha}{the p-value threshold for the dip test}

\item{data.name}{the name of the data to use in the output}
}
\value{
a list of data.frames, containing the columns of interest
}
\description{
Performs Hartigan's dip-test, and returns subset data frames containing
potentially non-unimodal columns plus the surrounding two columns.
}
\examples{
ap <- NMA_toy_dataset \%>\% dplyr::select(starts_with("Angle_profile"))
get.dip.test.regions(ap, "Angle_profile")

perims <- NMA_toy_dataset \%>\% dplyr::select(starts_with("Perimeter_microns"))
get.dip.test.regions(perims, "Perimeter")

}
