% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartigansdipper.R
\name{hartigansdipper}
\alias{hartigansdipper}
\title{non-unimodality detector(grouped)}
\usage{
hartigansdipper(dataset)
}
\arguments{
\item{dataset}{the dataset you want to take multimodal columns from
this function can itterate through the colunns of a dataset and add the
multimodal columns to a list in blocks of 5 this is designed aroud it being used
on regions of the profile regions of the dataset}
}
\description{
non-unimodality detector(grouped)
}
