% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graphs_for_bimodal_regions.R
\name{make.umapgraph}
\alias{make.umapgraph}
\title{Build a single umap chart over a bimodal region displaying clusters}
\usage{
make.umapgraph(clusters, umap, graphtype)
}
\arguments{
\item{clusters}{the clustered bimodal regions to be graphed}

\item{umap}{A single UMAP object}

\item{graphtype}{A string indicating the type of graph}
}
\description{
Build a single umap chart over a bimodal region displaying clusters
}
