% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graphs_for_bimodal_regions.R
\name{Make_profile_graphs}
\alias{Make_profile_graphs}
\title{Build a single profile chart over a bimodal region displaying clusters}
\usage{
Make_profile_graphs(profile_data, clusters, umaplist = umaplist, profiletype)
}
\arguments{
\item{profile_data}{the profile that will be graphed against}

\item{clusters}{the clustered bimodal regions to be graphed}

\item{umaplist}{A list of UMAP outputs}

\item{profiletype}{A string indicating the type of profile}
}
\description{
Build a single profile chart over a bimodal region displaying clusters
}
