% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_utility_functions.R
\name{MakeNMAConsensus}
\alias{MakeNMAConsensus}
\title{Generate a Consensus Polygon from Outline Data}
\usage{
MakeNMAConsensus(data, groups = NULL)
}
\arguments{
\item{data}{A data frame imported from NMA}

\item{groups}{Optional. A vector indicating group membership for each outline.}
}
\value{
A ggplot2 object showing the consensus outline(s)
}
\description{
This function computes a mean outline (consensus shape) from coordinate data.
If \code{groups} is provided, it computes one consensus outline per group.
}
