% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monohartigansdipper.R
\name{monohartigansdipper}
\alias{monohartigansdipper}
\title{Bimodality detector (single column)}
\usage{
monohartigansdipper(dataset)
}
\arguments{
\item{dataset}{the dataset you want to take bimodal columns from
this function can itterate through the colunns of a dataset and add the
bimodal columns to a list  this is designed aroud it being used
on regions that contain adjacent columns that aren't conected}

\item{selected_datasets}{the list you want to add bimodal columns to}
}
\description{
Bimodality detector (single column)
}
