% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_creation.R
\docType{data}
\name{NMA_Testdata}
\alias{NMA_Testdata}
\title{This is a dataset made from a Phodophus Cambelli, a fbxo7 mutant and a
PWK/lewes cross that has five different phenotypes in it}
\format{
An object of class \code{data.frame} with 6857 rows and 347 columns.
}
\usage{
NMA_Testdata
}
\description{
This is a dataset made from a Phodophus Cambelli, a fbxo7 mutant and a
PWK/lewes cross that has five different phenotypes in it
}
\keyword{datasets}
