% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NMA_toy_dataset}
\alias{NMA_toy_dataset}
\title{NMA toy dataset - A dataset With various phenotypes}
\format{
\subsection{"NMA_toy_dataset" a data frame with all the variables exported in}{

an NMA nuclear measurements file eg:Angle profiles, radius profiles and
circularity
}
}
\source{
SkinnerLab NMA data
}
\description{
This is a dataset made from a fbxo7 mutant that has at least two
different phenotypes in it. It is rounded to two decimals and sampled to save
space
}
