% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graphs_for_bimodal_regions.R
\name{plotbuilder3}
\alias{plotbuilder3}
\title{Build Graphs for Bimodal regions}
\usage{
plotbuilder3(
  clusters,
  originaldata,
  angle_data,
  diameter_data,
  radius_data,
  umaplist,
  selected_datasets
)
}
\description{
Build Graphs for Bimodal regions
}
