% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NMA_export.R
\name{read_NMA_export}
\alias{read_NMA_export}
\title{Read NMA Export File}
\usage{
read_NMA_export(path_to_export)
}
\arguments{
\item{path_to_export}{A string giving the file path to the tab-delimited NMA export file.}
}
\value{
A data frame containing the imported data.
}
\description{
Reads a tab-delimited nuclear morphology analysis (NMA) export file
into a data frame for analysis. The export must be produced using NMA with profiles and outlines included
for most package functionality.
}
\examples{
\dontrun{
data <- read_NMA_export("data/NMA_output.txt")
head(data)
}
}
