% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_profile_analysis.R
\name{targeted_profile_analysis}
\alias{targeted_profile_analysis}
\title{Targeted profile analysis
This function takes a NMA dataset and produces a page of graphs that help show
it's structure}
\usage{
targeted_profile_analysis(rawdata)
}
\arguments{
\item{rawdata}{the nuclear measurements exported dataset from NMA}
}
\description{
Targeted profile analysis
This function takes a NMA dataset and produces a page of graphs that help show
it's structure
}
