% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_profile_clusterer.R
\name{targeted_profile_clusterer}
\alias{targeted_profile_clusterer}
\title{clusters data automatically using hkmeans}
\usage{
targeted_profile_clusterer(selected_datasets, allow_further_itteration = FALSE)
}
\arguments{
\item{selected_datasets}{this is the output of the hartigansdippers and mono}

\item{allow_further_itteration}{if true increases clustering itterations
hartigans dippers.}
}
\description{
clusters data automatically using hkmeans
}
