% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartigansdipper.R
\name{get.dip.test.regions}
\alias{get.dip.test.regions}
\title{Get columns of a data.frame that are non-unimodal}
\usage{
get.dip.test.regions(data, dip.test.alpha = 0.05, is.profile = TRUE)
}
\arguments{
\item{data}{the data frame to input}

\item{dip.test.alpha}{the p-value threshold for the dip test}

\item{is.profile}{whether the region is a profile or not}
}
\value{
a list of data.frames, containing the columns of interest
}
\description{
Performs Hartigan's dip-test, and returns subset data frames containing
potentially non-unimodal columns plus the surrounding two columns.
}
