% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_profile_analysis.R
\name{targeted_profile_analysis}
\alias{targeted_profile_analysis}
\title{Targeted profile analysis
This function takes a NMA dataset and produces a page of graphs that help show
it's structure}
\usage{
targeted_profile_analysis(Data, verbose_output = FALSE)
}
\arguments{
\item{Data}{the file path to nuclear measurements exported dataset
from NMA}

\item{verbose_output}{gives the output as a list of graphs and data instead
of a Shiny interface}
}
\description{
Targeted profile analysis
This function takes a NMA dataset and produces a page of graphs that help show
it's structure
}
